memctl2
This is the initial commit of memctl2, a reimplementation of memctl that
will eventually become memctl. The idea is to redesign memctl from
scratch to be a post-exploitation framework for macOS and iOS.

Currently memctl2 has the basic skeleton of how the various pieces of
functionality will be implemented. However, nothing really works yet,
and there is a lot more to add before it becomes usable.