Implement kernel_call on AArch64
The default kernel_call implementation, kernel_call_7, has a number of
undesirable restrictions. kernel_call_aarch64 lifts these restrictions
on AArch64, allowing arbitrary kernel function calls with up to 8
arguments and with a 64-bit return value. In order to call kernel
functions a JOP payload is constructed and copied into the kernel, and
then the payload is executed using kernel_call_7.

This implementation is only preliminary. It uses hardcoded gadgets known
to exist on the iPhone 6s and iPhone 7 on iOS 10.1.1.