Implement safe memory read and f command
This commit adds the incredibly useful f (find) command to memctl,
supported by the ability to safely try to read arbitrary kernel
addresses. This implementation of kernel_read is not as powerful as the
implementation in the original memctl; it's a bit more conservative,
opting not to read any NUL mapping. Hopefully I'll find a better, more
flexible solution in the future. One consequence of this choice is that
the r command fails to read memory in the kernel image, since this
memory is inside a huge (4GB) NUL mapping.