Add vmp command to set virtual memory protection
This commit also fixes two issues in argument parsing. First, empty
strings in the argument list are now handled correctly. Second, if an
argument starts with a dash, it is sometimes ambiguous whether it is an
option or an argument. The command parser now treats such values as
potentially bad options, but defers reporting it as such until after the
argument parser has a chance to accept it.