Remove sandboxed function from kernel_call_7 setup
The previous implementation of kernel_call_init used the IOKit function
IORegistryEntryCreateCFProperty in order to discover the registry entry
ID of an IOUserClient it created. The IORegistryEntryCreateCFProperty
function is unavailable in some sandboxes, so kernel_call_init now looks
for an ID present after creation that was not present before creation of
the user client. No IOKit functions are used that were not present in
the old implementation, so this implementation is strictly more general.