Fix assertion failure during f command
An assertion in transfer_unsafe failed because the transfer_range_all
function returned with access set to 0 and when the first memory address
in the transfer range is unmapped. This issue arises in part because the
transfer function is still called when the size to transfer is 0. The
fix used here is to not call the transfer function if the transfer_range
function indicates that the region to be transferred has size 0. An
alternative fix would be to have transfer_range functions always return
with a valid access even if the size of the region to transfer is 0.