Add symbol finder for zone_element_size on iOS
The zone_element_size function is useful for finding the size of a
memory region allocated with kalloc. This change adds a special symbol
finder to locate this function by disassembling functions that call it.

The kernel symbol resolution functions now use the special symbol
finders in addition to standard symbol table lookup.