Improve initialization of kernel_call
There was a dependency issue between kernel_call and kernel_memory on
x86_64 that happened not to manifest when using the physmem core. The
x86_64 implementation of kernel_call depends on physical_write_unsafe to
overwrite the kernel TEXT, but physical_write_unsafe depends on
kernel_call to get the physical address to write and to perform the
write. This issue was addressed by making the kernel_call functions
return false if the corresponding implementation is not loaded and by
calling kernel_memory_init from within kernel_call_init_syscall_x86_64.