Allow specifying addresses as symbols in memctl
Any memctl command that takes an address will now also accept a symbol,
and the symbol is automatically converted into a virtual address.

This translation happens before options are interpreted, and so it is
not affected by options that specify the address should be interpreted
as a physical address instead.