Restructure kernel_call modules
This commit introduces two major changes: adding sizes to the arguments
to kernel_call, and factoring out the ARM64 JOP implementations into
separate files. Passing the size of the kernel function arguments to
kernel_call allows future kernel_call implementations to pass some
arguments to the kernel function on the stack, extending the maximum
number of arguments on supported platforms.