Add kcv command to call virtual methods
The kcv command just wraps the kc command to perform virtual method
calls (by index) on C++ objects in the kernel.

This commit also extends the 5-argument limit on the kc command by
adding a new argument type for arbitrary-length arrays of words.
However, this is not a true implementation of repeated arguments.